requirements.txt
pyusb
attrs
pandas
influxdb
python-daemon